refactor: 공통으로 사용하는 도메인 모델을 common 패키지로 추출
