+++ src/main/java/es/princip/getp/domain/project/domain/Project.java
@@ -1,7 +1,7 @@
 package es.princip.getp.domain.project.domain;
 
-import es.princip.getp.domain.base.BaseTimeEntity;
 import es.princip.getp.domain.client.domain.Client;
+import es.princip.getp.domain.common.domain.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
