+++ src/main/java/es/princip/getp/domain/project/domain/ProjectApplicationAttachmentFile.java
@@ -1,16 +1,7 @@
 package es.princip.getp.domain.project.domain;
 
-import es.princip.getp.domain.base.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import es.princip.getp.domain.common.domain.BaseTimeEntity;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
