+++ src/main/java/es/princip/getp/domain/project/dto/request/ApplicateProjectRequest.java
@@ -4,6 +4,7 @@ import es.princip.getp.domain.people.domain.People;
 import es.princip.getp.domain.project.domain.Project;
 import es.princip.getp.domain.project.domain.ProjectApplication;
 import jakarta.validation.constraints.NotNull;
+
 import java.time.LocalDate;
 import java.util.List;
