+++ src/main/java/es/princip/getp/domain/project/dto/response/CardProjectResponse.java
@@ -4,6 +4,7 @@ import es.princip.getp.domain.client.dto.response.ProjectClientResponse;
 import es.princip.getp.domain.project.domain.Project;
 import es.princip.getp.domain.project.domain.ProjectHashtag;
 import jakarta.validation.constraints.NotNull;
+
 import java.time.LocalDate;
 import java.util.List;
