+++ src/main/java/es/princip/getp/domain/serviceTerm/domain/ServiceTermAgreement.java
@@ -1,6 +1,6 @@
 package es.princip.getp.domain.serviceTerm.domain;
 
-import es.princip.getp.domain.base.BaseTimeEntity;
+import es.princip.getp.domain.common.domain.BaseTimeEntity;
 import es.princip.getp.domain.member.domain.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
